made it work with chapter
